Newcommers prep
A lot has been added on this commit. The getting stared dir was created and houses in progress instructions on how to download and install github and get it working. I have also added a template for script readmes that I need to beta test. I have also added numerous directories that will be relavent for adding scripts.